finished up forward phong lighting

+++ galaxy/src/galaxy/graphics/light/Point.cpp
@@ -23,73 +23,73 @@ namespace galaxy
 			switch (m_dist)
 			{
 				case Ranges::DIST_7:
-					m_linear    = 0.7;
-					m_quadratic = 1.8;
+					m_linear    = 0.7f;
+					m_quadratic = 1.8f;
 					break;
 
 				case Ranges::DIST_13:
-					m_linear    = 0.35;
-					m_quadratic = 0.44;
+					m_linear    = 0.35f;
+					m_quadratic = 0.44f;
 					break;
 
 				case Ranges::DIST_20:
-					m_linear    = 0.22;
-					m_quadratic = 0.20;
+					m_linear    = 0.22f;
+					m_quadratic = 0.20f;
 					break;
 
 				case Ranges::DIST_32:
-					m_linear    = 0.14;
-					m_quadratic = 0.07;
+					m_linear    = 0.14f;
+					m_quadratic = 0.07f;
 					break;
 
 				case Ranges::DIST_50:
-					m_linear    = 0.09;
-					m_quadratic = 0.032;
+					m_linear    = 0.09f;
+					m_quadratic = 0.032f;
 					break;
 
 				case Ranges::DIST_65:
-					m_linear    = 0.07;
-					m_quadratic = 0.017;
+					m_linear    = 0.07f;
+					m_quadratic = 0.017f;
 					break;
 
 				case Ranges::DIST_100:
-					m_linear    = 0.045;
-					m_quadratic = 0.0075;
+					m_linear    = 0.045f;
+					m_quadratic = 0.0075f;
 					break;
 
 				case Ranges::DIST_160:
-					m_linear    = 0.027;
-					m_quadratic = 0.0028;
+					m_linear    = 0.027f;
+					m_quadratic = 0.0028f;
 					break;
 
 				case Ranges::DIST_200:
-					m_linear    = 0.022;
-					m_quadratic = 0.0019;
+					m_linear    = 0.022f;
+					m_quadratic = 0.0019f;
 					break;
 
 				case Ranges::DIST_325:
-					m_linear    = 0.014;
-					m_quadratic = 0.0007;
+					m_linear    = 0.014f;
+					m_quadratic = 0.0007f;
 					break;
 
 				case Ranges::DIST_600:
-					m_linear    = 0.007;
-					m_quadratic = 0.0002;
+					m_linear    = 0.007f;
+					m_quadratic = 0.0002f;
 					break;
 
 				case Ranges::DIST_3250:
-					m_linear    = 0.0014;
-					m_quadratic = 0.000007;
+					m_linear    = 0.0014f;
+					m_quadratic = 0.000007f;
 					break;
 			}
 		}
 
-		const double Point::get_linear() const noexcept
+		const float Point::get_linear() const noexcept
 		{
 			return m_linear;
 		}
 
-		const double Point::get_quadratic() const noexcept
+		const float Point::get_quadratic() const noexcept
 		{
 			return m_quadratic;
 		}
